Refactor into new package 'task'
