+++ data/duke.txt
@@ -1,3 +1,4 @@
 T | 1 | Assignment
 E | 0 | Halloween Horror Night | 29/10/2023 | 31/10/2023
 D | 0 | Assignment | 30/1/2023 2359
+T | 0 | New Task
