+++ src/test/java/seedu/duke/DeadlineTest.java
@@ -1,6 +1,6 @@
 package seedu.duke;
 
-import duke.Deadline;
+import duke.task.Deadline;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
