+++ src/test/java/seedu/duke/EventTest.java
@@ -1,6 +1,6 @@
 package seedu.duke;
 
-import duke.Event;
+import duke.task.Event;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
