+++ src/test/java/seedu/duke/TodoTest.java
@@ -1,6 +1,6 @@
 package seedu.duke;
 
-import duke.Todo;
+import duke.task.Todo;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
